requirements.txt
fastapi==0.115.12
pydantic==2.11.4
docker==7.1.0
PyYAML==6.0.2
GitPython==3.1.44
uvicorn==0.34.2